requirements.txt
-i https://pypi.org/simple
aiohappyeyeballs==2.4.6; python_version >= '3.9'
aiohttp==3.11.13; python_version >= '3.9'
aiosignal==1.3.2; python_version >= '3.9'
amqp==5.3.1; python_version >= '3.6'
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.8.0; python_version >= '3.9'
attrs==25.1.0; python_version >= '3.8'
autoevals==0.0.120; python_full_version >= '3.8.0'
azure-ai-textanalytics==5.3.0; python_version >= '3.7'
azure-common==1.1.28
azure-core==1.32.0; python_version >= '3.8'
azure-identity==1.20.0; python_version >= '3.8'
azure-storage-blob==12.24.1; python_version >= '3.8'
beautifulsoup4==4.13.3; python_full_version >= '3.7.0'
billiard==4.2.1; python_version >= '3.7'
blinker==1.9.0; python_version >= '3.9'
blis==0.7.11; python_version < '3.13' and python_version >= '3.6'
boto3==1.37.0; python_version >= '3.8'
botocore==1.37.0; python_version >= '3.8'
braintrust==0.0.184; python_full_version >= '3.8.0'
braintrust-core==0.0.58; python_full_version >= '3.8.0'
bs4==0.0.2
catalogue==1.0.2; python_version >= '3.6'
celery==5.4.0; python_version >= '3.8'
certifi==2025.1.31; python_version >= '3.6'
cffi==1.17.1; platform_python_implementation != 'PyPy'
charset-normalizer==3.4.1; python_version >= '3.7'
chevron==0.14.0
click==8.1.8; python_version >= '3.7'
click-didyoumean==0.3.1; python_full_version >= '3.6.2'
click-plugins==1.1.1
click-repl==0.3.0; python_version >= '3.6'
cryptography==44.0.1; python_version >= '3.7' and python_full_version not in '3.9.0, 3.9.1'
cymem==2.0.11
cython==3.0.12; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
dataclasses-json==0.6.7; python_version >= '3.7' and python_version < '4.0'
distro==1.9.0; python_version >= '3.6'
exceptiongroup==1.2.2; python_version >= '3.7'
flask==3.1.0; python_version >= '3.9'
frozenlist==1.5.0; python_version >= '3.8'
gitdb==4.0.12; python_version >= '3.7'
gitpython==3.1.44; python_version >= '3.7'
googlemaps==4.10.0; python_version >= '3.5'
gunicorn==23.0.0; python_version >= '3.7'
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.7; python_version >= '3.8'
httpx==0.28.1; python_version >= '3.8'
httpx-sse==0.4.0; python_version >= '3.8'
idna==3.10; python_version >= '3.6'
isodate==0.7.2; python_version >= '3.7'
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.5; python_version >= '3.7'
jiter==0.8.2; python_version >= '3.8'
jmespath==1.0.1; python_version >= '3.7'
jsonpatch==1.33; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
jsonpointer==3.0.0; python_version >= '3.7'
jsonschema==4.23.0; python_version >= '3.8'
jsonschema-specifications==2024.10.1; python_version >= '3.9'
kombu==5.4.2; python_version >= '3.8'
langchain==0.3.19; python_version >= '3.9' and python_version < '4.0'
langchain-community==0.3.18; python_version >= '3.9' and python_version < '4.0'
langchain-core==0.3.39; python_version >= '3.9' and python_version < '4.0'
langchain-text-splitters==0.3.6; python_version >= '3.9' and python_version < '4.0'
langsmith==0.3.11; python_version >= '3.9' and python_version < '4.0'
levenshtein==0.26.1; python_version >= '3.9'
markupsafe==3.0.2; python_version >= '3.9'
marshmallow==3.26.1; python_version >= '3.9'
msal==1.31.1; python_version >= '3.7'
msal-extensions==1.2.0; python_version >= '3.7'
multidict==6.1.0; python_version >= '3.8'
murmurhash==1.0.12; python_version >= '3.6'
mypy-extensions==1.0.0; python_version >= '3.5'
numpy==1.26.4; python_version >= '3.9'
openai==1.64.0; python_version >= '3.8'
orjson==3.10.15; platform_python_implementation != 'PyPy'
packaging==24.2; python_version >= '3.8'
peewee==3.17.9
plac==1.1.3
portalocker==2.10.1; python_version >= '3.8'
preshed==3.0.9; python_version >= '3.6'
prompt-toolkit==3.0.50; python_full_version >= '3.8.0'
propcache==0.3.0; python_version >= '3.9'
pycparser==2.22; python_version >= '3.8'
pydantic==2.10.6; python_version >= '3.8'
pydantic-core==2.27.2; python_version >= '3.8'
pydantic-settings==2.8.0; python_version >= '3.8'
pyjwt[crypto]==2.10.1; python_version >= '3.9'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.1; python_version >= '3.8'
python-slugify==8.0.4; python_version >= '3.7'
pyyaml==6.0.2; python_version >= '3.8'
rapidfuzz==3.12.1; python_version >= '3.9'
redis==5.2.1; python_version >= '3.8'
referencing==0.36.2; python_version >= '3.9'
requests==2.32.3; python_version >= '3.8'
requests-toolbelt==1.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
rpds-py==0.23.1; python_version >= '3.9'
s3transfer==0.11.2; python_version >= '3.8'
setuptools==75.8.1; python_version >= '3.9'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
smmap==5.0.2; python_version >= '3.7'
sniffio==1.3.1; python_version >= '3.7'
soupsieve==2.6; python_version >= '3.8'
spacy==2.3.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
sqlalchemy==2.0.38; python_version >= '3.7'
srsly==1.0.7; python_version >= '3.6'
sseclient-py==1.8.0
tenacity==9.0.0; python_version >= '3.8'
text-unidecode==1.3
thinc==7.4.6; python_version < '3.13' and python_version >= '3.9'
tqdm==4.67.1; python_version >= '3.7'
typing-extensions==4.12.2; python_version >= '3.8'
typing-inspect==0.9.0
tzdata==2025.1; python_version >= '2'
urllib3==2.3.0; python_version >= '3.10'
vine==5.1.0; python_version >= '3.6'
wasabi==0.10.1; python_version >= '3.6'
wcwidth==0.2.13
werkzeug==3.1.3; python_version >= '3.9'
yarl==1.18.3; python_version >= '3.9'
zstandard==0.23.0; python_version >= '3.8'
fake-useragent
pygeocodio
usaddress
duckduckgo-search
langchain-openai